refactor: rename registry and base
