+++ aries_cloudagent/anoncreds/anoncreds/default/did_web/registry.py
@@ -12,7 +12,7 @@ from ....models.anoncreds_revocation import (
     AnonCredsRegistryGetRevocationRegistryDefinition,
 )
 from ....models.anoncreds_schema import GetSchemaResult
-from ...base_registry import BaseAnonCredsRegistrar, BaseAnonCredsResolver
+from ...base import BaseAnonCredsRegistrar, BaseAnonCredsResolver
 
 
 class DIDWebRegistry(BaseAnonCredsResolver, BaseAnonCredsRegistrar):
